Refactoring/bugfixes Renames notebooks to be ignored in documentation (Added _ in front)
