+++ water_security/data/model/model_best_params.csv
@@ -1,8 +1,10 @@
 ,risk0,risk1,risk2,risk3,risk4,risk5,risk6
-alpha,20.0,9.23904451238161,9.785299338867414,12.471391200643131,15.729827184320248,5.166255258956345,6.60153990724575
-colsample_bytree,0.3,0.6729732021314017,0.9,0.3,0.7983026006201641,0.5964385186176229,0.5198334908445289
-gamma,0.0,1.0,1.0,1.0,0.026171249246760175,0.9446855288305283,0.19920497769585288
-max_depth,4,1,1,1,5,1,4
-n_estimators,676,668,841,638,354,944,858
-subsample,0.8,0.8,0.8,0.8,0.8,0.8,0.8
+alpha,18.59764965453103,10.35207437772771,12.262000690691888,4.723972017517768,11.737668189343239,15.558190503443612,1.4761607081232042
+colsample_bytree,0.8356186094752351,0.6675003445305038,0.3881371433807168,0.9,0.8434239498560019,0.38256437803777865,0.6584855791736985
+gamma,0.806141513371056,0.8656113741339077,0.7876666851831892,1.0,0.3206497358973315,0.17246745965710897,0.558687939327392
+learning_rate,0.27542033787654746,0.46149368960622394,0.1247981864345916,0.5,0.2304958809444457,0.2676566850221501,0.4799779181770397
+max_depth,6,1,6,1,5,6,4
+n_estimators,639,638,955,582,590,306,391
+subsample,1,1,1,1,1,1,1
 objective,reg:squarederror,reg:squarederror,reg:squarederror,reg:squarederror,reg:squarederror,reg:squarederror,reg:squarederror
+random_state,42,42,42,42,42,42,42
